requirements.txt
Beaker==1.7.0
bottle==0.12.8
bottle-cork==0.11.1
gevent==1.0.2
greenlet==0.4.7
pycrypto==2.6.1
wsgiref==0.1.2
